general fibonacci implmenetation

+++ C/fibonacci.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 
-long fibonacci(int n) {
+long long fibonacci(int n) {
     return lround((
         pow(0.5 + 0.5 * sqrt(5.0), n)  -
         pow(0.5 - 0.5 * sqrt(5.0), n)) /
@@ -13,7 +13,7 @@ int main(void)
     int i, n=50;
     for (i=0; i<=n; i++)
         printf(
-            "Fibonacci of %02i is equal to %ld\n", i, fibonacci(i));
+            "Fibonacci of %02i is equal to %lld\n", i, fibonacci(i));
 
     return 0;
 }
